requirements.txt
pynput==1.8.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
PyWinCtl==0.4.1
PyRect==0.2.0
pytesseract==0.3.13
